Adding basic Gin server.
I'm not sure of the best way to manage dependencies. Seems like govendor is the right approach, but I couldn't get something I was fully satisfied with to make this easily re-installable. Instead I'm just documenting the two dependencies I'm using at the moment. I'll have to revisit that later when I get the chance.