Adding login page.
This allows the user to specify a username and then creates a User object for it and stores it at the top level. Only thing I couldn't get immediately working is that pressing "Enter" on the text entry does NOT submit the form. That's not a common or well-supported pattern on Android/iOS and I don't think Flutter Web has a great solution for that atm.

I'm also not a huge fan of the form validation recipe, but I did it in the spirit of writing Flutter-y code.