Adding RoomList screen.
This lists all rooms active on the server and allows the user to select one. Currently it just uses mock data, will implement on the server later.

Optionals were starting to get annoying, particularly for event listeners. I realized that an initializer list can be used to automatically make an input nullable. This provides the same compile-time guarantees for the widget implementation, but allows its users to not worry about it because simply providing the event listener or not will "just work".