Adding tests for non-Flutter code.
I've been a bad boy and wasn't writing tests for these changes. Started off with some tests for the non-Flutter code to minimize dependencies. None of these tests are very interesting, it's just good to have them there so they can be easily added to when interesting changes are introduced.

I contemplated a makefile to handle build/run/test operations, but I don't see it reducing much complexity. I can't give it obvious names like `make client` because the client/ directory exists and thus always returns "make: 'client' is up to date." Weird names would make it just as awkward to use as well. Flutter's tooling doesn't like being in the non-root directory, so this doesn't work out very cleanly. Maybe the solution is to separate the server and client into different repositories? I'll have to brew on that a little bit.