Investigated testing for Flutter Web.
Long story short, Flutter Web's tech preview doesn't seem to support testing very well. I suspect it would be possible to hack something together, but I'm not sure it's worth the effort at the moment for a small project like this.

Instead, I'm just documenting that this doesn't seem to be properly supported at the moment and I'm not going to bother with it.