Switching to Go modules.
Came across the new modules system which removes all the GOPATH weirdness and makes the directory structure much simpler.

I opted to created the module at the root directory rather than the server/ directory. This is to avoid an unncessary `cd` command before building/running and because Go code could (hypothetically) be used outside the server, so including "server/" in the import paths would better handle that case.

I couldn't get local imports to work, so instead all imports are fully-qualified. I think that's the way it's supposed to work, but I couldn't find any direct confirmation of that.