Connecting client room list to server.
Used http.Client for the network requests. There seem to be multiple official Dart/Flutter solutions for network requests, but this was the only one which 1) worked with Flutter Web and 2) allowed me to inject a client which could be mocked in tests. I'm dissapointed that Dart doesn't support spies which would make testing a lot easier and remove the need to inject dependencies like this.

I also came up with the idea of FutureValue(), which is a simple Flutter widget that takes a Future with some callbacks to render that Future in all three of its major states (pending, completed, errored). This is pretty useful for rendering async operations to a synchronous UI. If I get some time, I might separate that out to a Flutter library for others to take advantage of.

I also had to update Room to override == just so it could be asserted with. Longer term solution might be to use some kind of built-value codegen rather than managing this manually.