Adding alert dialog and retry button on connection error.
I tried adding an amusing animated gif on error, but I could not get the gif to play. After some initial debugging, I tried numerous ways of creating the image (Image.network(...), Image(image: AssertImage(...)), couple others...), but none of them would animate the image. I suspect this might be a Flutter Web bug, but I'm too lazy to chase it down right now.

I also discovered that FutureValue does not seem to handle the Future changing. for some reason, initState() is not called when the `future` attribute is changed in a new build() call. I'm not sure why that would not be re-initialized, but it means that retrying the API call would not update the UI.

I changed this over to use the standard FutureBuilder. I find the API at lot more complex, but it updates correctly which is the most important thing.