Adding test Docker images.
These images run the client and server tests in a hermetic fashion.

I'm hoping to use these to run CI tests in an automated fashion. Unfortunately, Docker doesn't seem to support this very well atm. This seems to be somewhat possible with docker-compose and `autotest` in Docker Hub. However, I can't figure out how to get it to use a particular stage for the test. You can select different Dockerfiles, but then I'd need to re-use the existing Dockerfile for the main build, which doesn't appear to be well supported either without significant hacking.