Upgrading Gradle and Docker builds.
This updates the version numbers of everything to be more modern. It also replaces ShadowJar with a manual operation as I couldn't get ShadowJar to work with newer versions.

I also separated the dependency download step from the build step in Docker, so we can skip the former when making regular source changes and improve iterative build times.