Switching frontend server to TypeScript/Node.
I decided against Kotlin becuase it did not play well with VSCode. It seems there is no official language support for Kotlin in VSCode (probably because JetBrains owns both Kotlin and IntelliJ). I didn't want to swap between multiple editors, so I looked for other languages which might work better.

I considered Rust, but found that the monorepo gave some challenges with Rust workspaces and wasn't able to get VSCode editor integration working to my liking. Also Rust doesn't support Docker well and there is no way to pre-cache dependencies, which heavily inflates build times.

In the end, I just went with Node/TypeScript, which is something I'm more familiar with. It wasn't too hard to get it working with Docker by simply running `tsc`, however this might cause problems later on as I don't have a config file (which would confuse vscode). This also ends up mixing Dart on the frontend with TypeScript on the backend, which is kind of a weird combination. However, I'm not using Dart on the frontend because I like Dart, rather I'm using it to take advantage of Flutter. I don't anticipate opportunities to share code between the frontend and backend anyways, so I don't see too much of an issue using TypeScript on the backend.