Adding rooms microservice.
This is currently just a "Hello world" service. Mostly a clone of frontend service's Docker build.

Unfortunately I had to duplicate the `tsc` Docker container as there's no good way to "import" one Dockerfile from another. I'd have to publish the `tsc` container in order to load it as an import. Mayble I'll do that in the future, but for now it's a small enough duplication that I'm not too concerned with it.