Adding simpleResponder() as a means of easily testing Express route handlers.
This returns the response rather than editing a Response parameter, which is easier to test because the result can just be asserted rather than spying for side effects on the provided Response parameter.

Currently, these utilities are somewhat trapped in the rooms service module and not easily sharable across other microservices. I don't see a good way of doing that in this monorepo with Docker, since that's not really how Dockerfiles are supposed to work. I think the best way of doing that would be to make this a separate Node module, publish it to NPM, and then install it that way, rather than trying to share code with Docker. I'm not bothering with that just yet, but will likely need to do that to easily reuse this code across other microservices.