Adding `/api/rooms/create` API.
Pretty straightforward implementation, though a lot of little things needed to happen to make it so simple.

1) tsconfig.json outputs as ES6 so `instanceof` works as expected, but uses CommonJS so imports work in Node.
2) Jasmine config looks for any files ending in "*_test.js" to run those tests.
3) Previous commits included `simpleResponder()` which enables this commit to test the API much easier.

Oddly, the `Room` model is responsible for serailizing itself, but not deserializing itself. This is because the ID value does not come from user input, but rather from outside context. So a `parse()` function is not as self-contained as it might appear at first glance.