Making TypeScript project strict.
I thought TS was strict by default, but was incorrect. I added the "strict" flag to the config to auto-enable the simple strict configuration. I might tweak the flags later as appropriate, but otherwise this should be fine as it includes `noImplicitAny` and `strictNullChecks` which are the two most important IMHO.

This did reveal that working with `unknown` is a little more restrictive than I thought, but I was able to get it working without too much fiddling.