Adding `/api/rooms/list` API.
This is pretty straightforward since it is just listing out all the rooms which are stored in-memory.

Only interesting part is that I changed the Room class' `stringify()` function to `serialize()` and use a plain JavaScript object rather than a JSON string. This makes it more composable for use-cases like `list` where several Room objects need to be serialized as a single list.