Switches Kubernetes config to use new microservices.
This switches off  the old monolithic server and uses the new microservices with an ingress for path routing.

I was able to get this to work with Minikube and documented the steps to do it. I haven't tested this on GCP just yet, but I think this is a good step already. There are still a few things I'm not totally happy with:

1) When testing locally you need to manually add `imagePullPolicy: Never` to the deployments or else it won't use local images. I think longer term, I'll have prod/staging/dev environments, and the dev can just be for local use with that flag included, while prod and staging would not.

2) I figured out that the Docker issues I was having with the frontend server are because I was in a Minikube context. I worked around this by building on my host machine and then pushing it to the Minikube instance. This is harder than you'd think, because the registry is embedded in  the Docker tag, which also needs to be hard coded in the Kubernetes YAML config. I found the ability to serialize a Docker image to a tar and push it that way, which is a surprsingly elegant way of handling this problem.

3) The ingress configuration is not nearly as usable as I wish it was. For some reason the `rewrite-target` annotation appears to be required, despite the fact that I don't need it. A host is also required, which makes local debugging tricky since we can't use the IP directly. The way that it resolves `serviceName` values is a little confusing to me as well. This commit was a massive pain of debugging and took way longer than it should have, largely for these reasons.

I also updated the documentation to explain how I got this working on Minikube. I need to generally update the docs as well and add service-specific docs in the coming commits.