Updating Kubernetes config to work with Google Kubernetes Engine.
After some fiddling I was eventually able to get the Ingress to work on GKE. I had a lot of difficulty getting the static IP to hold, and I'm not convinced the global IP annotation is working correctly, as I could not get it to use the existing IP becuase it would always use a new one it generated. I instead reserved that new one as a static IP and it seems to be picked up consistently on restarts. I'm not entirely convinced that this is loading the static ip correctly, it might just be resuming the last one it had on restart. Either way, this is working for now and I can worry about it later if it suddenly stops working.

Secondly, I spent some time trying to get the domain name to update the IP it pointed to. I kept trying to change Google Domains to point to the new IP, but it never seemed to take. Eventually, I discovered that Google Domains doesn't work that way with GKE. Instead, I needed to point it to Google Cloud's nameservers and then configure them within Google Cloud Console. For whatever reason, Google Domains still allows you to edit the UI which won't do anything because it's pointed to foreign nameserver. I added a record for "dont.use.this.ui.use.gcloud.instead.chatter.technology" to try and remind myself of this the next time I need to change it.

After all this, visiting http://chatter.technology/ finally works using the two micro services. The frontend services correctly and the API layer works as expected. The ingress correctly routes the different URLs to the right services. I also upgraded Kubernetes to the currently latest version and lowered the number of nodes to just 1 to try and save on billing a little bit since up-time isn't as important as lowering costs for a toy service like this.