Adds persistent volumes to Redis instance.
This binds the Redis instance to a persistent volume. Tested by running this in Minikube, port forwarding the Redis instance and connecting to it with a CLI. I incremented a counter and saved the data, then forcibly killed the instance and restarted it. After reconnecting to the instance I found the counter was still present.

I believe GKE will work differently as it appears to manage volumes directly, much like ingress IPs. I'm saving that for a separate commit and will likely have to diverge the configurations between dev and prod.