Connects RoomsService to Redis data store.
This commit got a little larger than initially expected, and I'm being a little lazy by not splitting it up. Quite a few things needed to happen to get this to work correctly.

I used the `redis` NPM module to interact with the backend database. However, its API is all callback based and I wasn't satisfied with it. Instead, I wrapped the API in a Promise-based version which works a lot better.

I also used `redis-mock` and wrapped it in the same manner for testing purposes. Unfortunately there is no supported way of stubbing the clock, which makes key expiration tricky to test. I did my best for this commit and filed to get a proper solution: https://github.com/yeahoffline/redis-mock/issues/127.

During implementation, I also upgraded to Node 11 and set `tsc` to output ES2019. This let me use `Array.prototype.flatMap()` and numeric separators. I also added a "debug test" NPM script and Docker target with some documentation about how to effectively debug a broken test. I'm not totally satisified with this developer experience, but I'll need to revisit it when I have more time in the future.