Add specific error message when making an `/api/rooms/create` call without `application/json` Content-Type.
Express requires this specific header to parse the JSON, without it the request will fail. However, this has given a poor error message making the Content-Type requirement hard to understand. This commit adds a special error message to explicitly call out such a mistake when made.

Also created a `RequestFake` class to support this requirement as Express `Request` objects are not easily mockable for their headers.