Uploads Docker image as an artifact
The previous run failed presumably because jobs do not implicitly share state, so the built Docker image was not available to the `publish` job. https://github.com/dgp1130/chatter/actions/runs/36443345

Updated the `build` job to save the image to a TAR file and upload it as an artifact after build. Then the `publish` job downloads that artifact and loads it before attempting to publish to the registry.