Moves login functionality into `publish` job.
Since data is not persisted between jobs, the `login` job is storing credentials which don't get propagated to the `publish` job.

I could have Docker write them to a particular config file path and then upload that as an artifact. However I'm not sure of the security of GitHub artifacts to know if it is reasonable to put a secret inside there.

Requiring users of the `login` job to explicitly download the artifact is also a pain. May as well just login as part of the job.