Adds test step to workflow.
This builds and runs the test image before publishing to Docker Hub. It's a little awkward since the test image is actually a different build. Since both the test and release image come from the same commit, we have reasonably good assertions that they will be consistent. (They would not only if there is a non-hermetic build or Docker itself encounters some bug.)