Adds async render library module.
This defines a set of directives which render all states of an `Observable`. It is effectively a more feature-complete replacement for `AsyncPipe`, which does not handle error or complete events from an `Observable`. This allows an `Observable` to be used directly in the view, while requiring the view to handle possible errors (and optionally handle a completed state).