Adds /api/deployed to Rooms service.
Last deployment workflow failed because the service account used doesn't have the right permissions. I updated it in GKE to use "Project Editor" permissions, which is hopefully generic enough to provide everything it needs.

This commit is just to add a simple endpoint to the Rooms service to verify that the rollout is successful when the workflow completes. This will also trigger another run to test the updated permissions.