Sets `revisionHistoryLimit` and CPU resources.
Pods were not able to be scaled because there was not enough CPU resources on the single node I'm paying for. By default, each deployment requires 0.1 "CPU", of which the node only provides 1 CPU and most of that goes to internal Kubernetes services.

I removed the CPU requirements for Chatter services since they are all relatively trivial and receive ~0 QPS anyways.

Also set `revisionHistoryLimit` to 2 to limit the amount of empty deployments. No need to be able to roll back the default 10 iterations. I figure a rollback of 2 is good enough for a service like this.