Removes references to running GKE instance.
I just tore down the GCP project and dropped the domain name because I'm cheap and haven't been working on this project for some time now. The main things I want to work on in the future is becoming more familiar with Angular, which doesn't require an enterprise-scale production environment or the costs that come with it. If I ever feel like playing more with the productionization side, then I might revive this at some point in the future.